break `sw_cond_add` into `cond_add` and `sw_cond_add`
