+++ common/src/gadgets/mod.rs
@@ -5,6 +5,7 @@ use ark_std::vec::Vec;
 
 pub mod booleanity;
 // pub mod inner_prod_pub;
+pub mod cond_add;
 pub mod sw_cond_add;
 pub mod fixed_cells;
 pub mod inner_prod;
